+++ resources/js/Pages/Users/{$id}/Edit.vue
@@ -27,7 +27,7 @@
                 <div>
                     <div class="mb-6"><img :src="props.userEdit.profile_photo_url" class="rounded-full h-20 w-20 object-cover"/></div>
                     <div class=""><span class="text-xs uppercase">User ID: </span><span class="font-semibold">{{props.userEdit.id}}</span></div>
-                    <div class=""><span class="text-xs uppercase">Subscription Status: </span><span class="font-semibold">{{props.userEdit.subscriptionStatus}}</span></div>
+                    <div class=""><span class="text-xs uppercase">Subscription Status: </span><span class="font-semibold">{{props.subscriptionStatus}}</span></div>
                     <div class="" v-if="props.userEdit.role_id == 4"><span class="text-xs uppercase">Creator #: </span><span class="font-semibold">{{props.userEdit.creatorNumber}}</span></div>
                 </div>
                 <div class="flex flex-col space-y-2 align-bottom">
@@ -276,7 +276,7 @@ let props = defineProps({
     isVip: Boolean,
     isSubscriber: Boolean,
     hasSubscription: null,
-    subscriptionStatus: '',
+    subscriptionStatus: String,
 });
 
 let form = useForm({